+++ src/common/utils/get_resolved_document_by_id.py
@@ -2,7 +2,7 @@ from typing import Callable
 
 from common.address import Address
 from common.exceptions import ApplicationException
-from common.utils.is_reference import is_reference
+from common.utils.data_structure.is_reference import is_reference
 from common.utils.resolve_address import ResolvedAddress, resolve_address
 from storage.data_source_class import DataSource
