+++ src/features/document/use_cases/get_document_use_case.py
@@ -2,8 +2,8 @@ from pydantic import conint
 
 from authentication.models import User
 from common.address import Address
-from common.tree_node_serializer import tree_node_to_dict
-from domain_classes.tree_node import ListNode, Node
+from common.tree.tree_node import ListNode, Node
+from common.tree.tree_node_serializer import tree_node_to_dict
 from services.document_service import DocumentService
 from storage.internal.data_source_repository import get_data_source
