+++ src/tests/unit/services/document_service/test_get_extended_blueprint.py
@@ -2,9 +2,9 @@ import unittest
 from unittest import mock
 
 from common.address import Address
+from common.tree.tree_node import Node
 from common.utils.data_structure.compare import get_and_print_diff
 from domain_classes.blueprint import Blueprint
-from domain_classes.tree_node import Node
 from tests.unit.mock_data.mock_blueprint_provider import MockBlueprintProvider
 from tests.unit.mock_data.mock_document_service import get_mock_document_service
