+++ src/tests/unit/services/document_service/test_reference_resolve.py
@@ -5,7 +5,7 @@ from unittest import mock
 import pytest
 
 from common.address import Address
-from common.tree_node_serializer import tree_node_to_dict
+from common.tree.tree_node_serializer import tree_node_to_dict
 from common.utils.data_structure.compare import get_and_print_diff
 from common.utils.data_structure.has_key_value_pairs import has_key_value_pairs
 from enums import REFERENCE_TYPES, SIMOS, Protocols
