+++ src/tests/unit/services/document_service/test_save.py
@@ -4,9 +4,9 @@ from unittest import mock
 
 from authentication.models import User
 from common.address import Address
+from common.tree.tree_node import Node
 from common.utils.data_structure.compare import get_and_print_diff
 from domain_classes.blueprint_attribute import BlueprintAttribute
-from domain_classes.tree_node import Node
 from enums import REFERENCE_TYPES, SIMOS
 from features.document.use_cases.update_document_use_case import (
     update_document_use_case,
