+++ src/tests/unit/tree_functionality/test_tree_node_from_dict.py
@@ -1,7 +1,7 @@
 import json
 import unittest
 
-from common.tree_node_serializer import tree_node_from_dict, tree_node_to_dict
+from common.tree.tree_node_serializer import tree_node_from_dict, tree_node_to_dict
 from enums import REFERENCE_TYPES, SIMOS
 from tests.unit.mock_data.mock_blueprint_provider import MockBlueprintProvider
