+++ src/tests/unit/tree_functionality/test_tree_node_to_ref_dict.py
@@ -1,8 +1,8 @@
 import unittest
 
-from common.tree_node_serializer import tree_node_to_ref_dict
+from common.tree.tree_node import ListNode, Node
+from common.tree.tree_node_serializer import tree_node_to_ref_dict
 from domain_classes.blueprint_attribute import BlueprintAttribute
-from domain_classes.tree_node import ListNode, Node
 from enums import REFERENCE_TYPES, SIMOS
 from tests.unit.mock_data.mock_blueprint_provider import MockBlueprintProvider
 from tests.unit.tree_functionality.mock_data_for_tree_tests.get_node_for_tree_tests import (
